Add abstract viewmodel test for main thread execution
